Refactor leaf select/deselect code
